+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/tests/setup.js'],
+  testMatch: [
+    '<rootDir>/tests/**/*.test.js',
+    '<rootDir>/tests/**/*.spec.js'
+  ],
+  collectCoverageFrom: [
+    'src/js/**/*.js',
+    '!src/js/**/*.test.js',
+    '!src/js/**/*.spec.js'
+  ],
+  coverageDirectory: 'coverage',
+  coverageReporters: ['text', 'lcov', 'html']
+};